[server] Remove Workspace Factory EE, merge with standard
